Split LoadBalancerProvider and the registry; service loader works
