+++ core/src/main/resources/META-INF/services/io.grpc.LoadBalancerProvider
@@ -1,0 +1,2 @@
+io.grpc.internal.PickFirstLoadBalancerProvider
+io.grpc.internal.RoundRobinLoadBalancerProvider
