+++ core/src/test/java/io/grpc/internal/RoundRobinLoadBalancerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.util;
+package io.grpc.internal;
 
 import static com.google.common.truth.Truth.assertThat;
 import static io.grpc.ConnectivityState.CONNECTING;
@@ -22,7 +22,7 @@ import static io.grpc.ConnectivityState.IDLE;
 import static io.grpc.ConnectivityState.READY;
 import static io.grpc.ConnectivityState.SHUTDOWN;
 import static io.grpc.ConnectivityState.TRANSIENT_FAILURE;
-import static io.grpc.util.RoundRobinLoadBalancerFactory.RoundRobinLoadBalancer.STATE_INFO;
+import static io.grpc.internal.RoundRobinLoadBalancer.STATE_INFO;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
@@ -60,11 +60,10 @@ import io.grpc.Metadata;
 import io.grpc.Metadata.Key;
 import io.grpc.Status;
 import io.grpc.internal.GrpcAttributes;
-import io.grpc.util.RoundRobinLoadBalancerFactory.EmptyPicker;
-import io.grpc.util.RoundRobinLoadBalancerFactory.ReadyPicker;
-import io.grpc.util.RoundRobinLoadBalancerFactory.Ref;
-import io.grpc.util.RoundRobinLoadBalancerFactory.RoundRobinLoadBalancer;
-import io.grpc.util.RoundRobinLoadBalancerFactory.RoundRobinLoadBalancer.StickinessState;
+import io.grpc.internal.RoundRobinLoadBalancer.EmptyPicker;
+import io.grpc.internal.RoundRobinLoadBalancer.ReadyPicker;
+import io.grpc.internal.RoundRobinLoadBalancer.Ref;
+import io.grpc.internal.RoundRobinLoadBalancer.StickinessState;
 import java.net.SocketAddress;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -86,7 +85,7 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 
-/** Unit test for {@link RoundRobinLoadBalancerFactory}. */
+/** Unit test for {@link RoundRobinLoadBalancer}. */
 @RunWith(JUnit4.class)
 public class RoundRobinLoadBalancerTest {
   private RoundRobinLoadBalancer loadBalancer;
@@ -132,8 +131,7 @@ public class RoundRobinLoadBalancerTest {
           }
         });
 
-    loadBalancer = (RoundRobinLoadBalancer) RoundRobinLoadBalancerFactory.getInstance()
-        .newLoadBalancer(mockHelper);
+    loadBalancer = new RoundRobinLoadBalancer(mockHelper);
   }
 
   @After
